requirements.txt
behave==1.2.6
selenium==3.141.0
webdriver-manager==3.3.0
Appium-Python-Client~=1.1.0